+++ CHANGELOG.md
@@ -8,6 +8,13 @@ All notable changes to this project will be documented in this file.
 ### 🚀 Features
 
 - Add FileDownloadCrawler for downloading and streaming files ([#2043](https://github.com/apify/crawlee-python/pull/2043)) ([90d550d](https://github.com/apify/crawlee-python/commit/90d550d8a4b4dfe134c76086a5390b7ecea1d1a7)) by [@Mantisus](https://github.com/Mantisus), closes [#875](https://github.com/apify/crawlee-python/issues/875)
+- **dataset:** Add `collect_all_keys` option and keep CSV export columns aligned ([#2079](https://github.com/apify/crawlee-python/pull/2079)) ([cd1e1f6](https://github.com/apify/crawlee-python/commit/cd1e1f632cab692a54ba8db63e9f64ce3130a6d6)) by [@anxkhn](https://github.com/anxkhn)
+
+### 🐛 Bug Fixes
+
+- **cli:** Migrate CLI to `typer` &gt;=0.26 ([#2080](https://github.com/apify/crawlee-python/pull/2080)) ([c864b96](https://github.com/apify/crawlee-python/commit/c864b96d71e18649ab34e18cc6303beee2ff44e9)) by [@Mantisus](https://github.com/Mantisus), closes [#2063](https://github.com/apify/crawlee-python/issues/2063)
+- **system:** Do not fail memory snapshots on inaccessible processes ([#2078](https://github.com/apify/crawlee-python/pull/2078)) ([5ad4e53](https://github.com/apify/crawlee-python/commit/5ad4e53ba9bfbaa0147dd9535bfcadf0c89ce3ff)) by [@anxkhn](https://github.com/anxkhn)
+- **memory-storage:** Stop re-adds from duplicating and resetting pending requests ([#2074](https://github.com/apify/crawlee-python/pull/2074)) ([fae204b](https://github.com/apify/crawlee-python/commit/fae204bc2c4557072fe9e13c3ec0cfaf49aa4717)) by [@anxkhn](https://github.com/anxkhn)
 
 
 <!-- git-cliff-unreleased-end -->